Split wiring into multiple files
