+++ app/wiring/ApplicationConfiguration.scala
@@ -1,0 +1,7 @@
+package wiring
+
+import config.Configuration
+
+trait ApplicationConfiguration {
+  val appConfig = new Configuration()
+}
